Refactor, move render functions into NebariPluginManager
