+++ src/_nebari/subcommands/deploy.py
@@ -3,8 +3,8 @@ import pathlib
 import typer
 
 from _nebari.deploy import deploy_configuration
-from _nebari.render import render_template
-from nebari import schema
+
+# from _nebari.render import render_template
 from nebari.hookspecs import hookimpl
 
 
@@ -62,12 +62,10 @@ def nebari_subcommand(cli: typer.Typer):
         from nebari.plugins import nebari_plugin_manager
 
         stages = nebari_plugin_manager.ordered_stages
-        config_schema = nebari_plugin_manager.config_schema
+        config = nebari_plugin_manager.read_configuration(config_filename)
 
-        config = schema.read_configuration(config_filename, config_schema=config_schema)
-
-        if not disable_render:
-            render_template(output_directory, config, stages)
+        # if not disable_render:
+        #     render_template(output_directory, config, stages)
 
         deploy_configuration(
             config,
