+++ src/_nebari/subcommands/validate.py
@@ -3,7 +3,6 @@ import pathlib
 import typer
 from rich import print
 
-from nebari import schema
 from nebari.hookspecs import hookimpl
 
 
@@ -31,7 +30,6 @@ def nebari_subcommand(cli: typer.Typer):
         else:
             from nebari.plugins import nebari_plugin_manager
 
-            schema.read_configuration(
-                config_filename, config_schema=nebari_plugin_manager.config_schema
-            )
+            nebari_plugin_manager.read_configuration(config_filename)
+
             print("[bold purple]Successfully validated configuration.[/bold purple]")
