+++ src/nebari/schema.py
@@ -245,46 +245,3 @@ def set_config_from_environment_variables(
             )
             sys.exit(1)
     return config
-
-
-def read_configuration(
-    config_filename: pathlib.Path,
-    read_environment: bool = True,
-    config_schema: pydantic.BaseModel = Main,
-):
-    """Read configuration from multiple sources and apply validation"""
-    filename = pathlib.Path(config_filename)
-
-    yaml = YAML()
-    yaml.preserve_quotes = True
-    yaml.default_flow_style = False
-
-    if not filename.is_file():
-        raise ValueError(
-            f"passed in configuration filename={config_filename} does not exist"
-        )
-
-    with filename.open() as f:
-        config = config_schema(**yaml.load(f.read()))
-
-    if read_environment:
-        config = set_config_from_environment_variables(config)
-
-    return config
-
-
-def write_configuration(
-    config_filename: pathlib.Path,
-    config: typing.Union[Main, typing.Dict],
-    mode: str = "w",
-    config_schema: pydantic.BaseModel = Main,
-):
-    yaml = YAML()
-    yaml.preserve_quotes = True
-    yaml.default_flow_style = False
-
-    with config_filename.open(mode) as f:
-        if isinstance(config, config_schema):
-            yaml.dump(config.dict(), f)
-        else:
-            yaml.dump(config, f)
